refactor: js파일 js폴더에서 관리/ simplify: mainEvent.js 데이터 받아서 렌더 함수 호출하게 역할 단일/ view폴더 파일 이름 변경 및 함수 내용 폴더에 맞게 조정
