+++ public/index.js
@@ -1,12 +1,4 @@
-import { domUtil } from "./Util/util.js";
-
-import {
-  renderMainSecHome,
-  renderMainSecWoD,
-  renderBanner,
-} from "./eventJS/recEvent.js";
-import { data } from "./data/data.js";
-import { init } from "./Util/init.js";
+import { init } from "./js/Util/init.js";
 let WHOLE_DATA;
 
 fetch("/index/wholeData")
