+++ public/js/component/BannerComponent.js
