+++ public/js/component/mainComponent.js
